+++ src/templateFunction.hpp
@@ -4,9 +4,9 @@
 // E stands for Enum. 
 template <class D , typename E>
 // Pushes <total> times , <D type> objects in the given <deck>
-void pushNtimes(std::shared_ptr < std::queue< std::shared_ptr < D > > > deck , std::unordered_map<std::string , std::vector<size_t> >::iterator j ,const E type  , size_t total , size_t * times ) {
+void pushNtimes(std::shared_ptr < std::queue< std::shared_ptr < D > > > deck , std::unordered_map<std::string , std::vector<uint16_t> >::iterator j ,const E type  , uint16_t total , uint16_t * times ) {
 
-  size_t temp = total;
+  uint16_t temp = total;
   while (temp--)
   {
     if (j->second.size() == 6) {  // Follower
